requirements.txt
cmake==3.29.2
dlib @ file:///Users/runner/miniforge3/conda-bld/dlib-split_1712214080496/work
imutils==0.5.4
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225416355/work/dist/numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=14c628943f67e9c10d22faa9c1cdcbf69c3192e88f2f597d03b77a252e589567
opencv-python==4.9.0.80
pyserial==3.5
